+++ PdfMerger/Classes/RandomPassword.cs
@@ -1,0 +1,22 @@
+﻿using System.Security.Cryptography;
+
+namespace PdfMerger.Classes;
+
+
+public static class RandomPassword
+{
+    private const string chars = "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnpqrstuvwxyz23456789";
+
+    public static string Generate(int length)
+    {
+        var data = RandomNumberGenerator.GetBytes(length);
+        var result = new char[length];
+
+        for (int i = 0; i < length; i++)
+        {
+            result[i] = chars[data[i] % chars.Length];
+        }
+
+        return new string(result);
+    }
+}