+++ PdfMerger/Properties/Resources.Designer.cs
@@ -290,6 +290,16 @@ namespace PdfMerger.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
+        internal static System.Drawing.Bitmap security {
+            get {
+                object obj = ResourceManager.GetObject("security", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Apache License
         ///Version 2.0, January 2004
         ///http://www.apache.org/licenses/